refactor: 앨범에 저장하는 로직 삭제, CoreData의 마지막 데이터 썸네일 불러오기 로직 수정
